+++ scheduler/run_trace.py
@@ -8,18 +8,21 @@ import argparse
 
 def main():
     parser = argparse.ArgumentParser()
-    parser.add_argument('--name_test', type=str, required=True, help='name of test (required)')
-    parser.add_argument('--policies', type=list, default=["fcfs", "priority"], help='policies executed (fcfs and/or priority)')
-    parser.add_argument('--assigment_policies', type=list, default=["strict", "always_attend", "max_prop"], help= 'resource allocation strategies that were used (strict and/or always_attend and/or max_prop')
-    parser.add_argument('--tf_version', type=list, default=["maleable" , "original"], help= 'tensorflow versions used (malleable and/or original)')
+    parser.add_argument('-n', '--name_test', type=str, required=True, help='name of test (required)')
+    parser.add_argument('-p', '--policies', type=str, default="fcfs priority", help='policies executed (fcfs and/or priority)')
+    parser.add_argument('-a', '--assigment_policies', type=str, default="strict always_attend max_prop", help= 'resource allocation strategies that were used (strict and/or always_attend and/or max_prop')
+    parser.add_argument('-t', '--tf_version', type=str, default="maleable original", help= 'tensorflow versions used (malleable and/or original)')
     args = parser.parse_args()
     name_test= args.name_test
+    policies= args.policies.split(' ')
+    assigment_policies= args.assigment_policies.split(' ')
+    tf_versions= args.tf_version.split(' ')
     f = open(name_test+'.csv', "w")
-    for policy in args.policies:
+    for policy in policies:
         n_containers=8
         for i in range(5):
-            for assigment_policy in args.assigment_policies:
-                for tf_version in args.tf_version:
+            for assigment_policy in assigment_policies:
+                for tf_version in tf_versions:
                     run_trace=  "python3 Scheduler-host/Commons/trace.py Data/log/" + name_test + "/" + policy + "/" + assigment_policy + "/" +  str(n_containers) + "Contenedores_" + tf_version + "/" + " gantt_events.txt " + str(n_containers)
                     print(run_trace)
                     process_command = Popen([run_trace], stdout=PIPE, stderr=PIPE, universal_newlines=True, shell=True)